Port tokenizer to Typescript
This ports over the tokenizer from elasticlunr.js. It takes in either a
single string or an array of strings and returns an array of tokens from
it. It will split the string on the separator that it has, which will
either be the default one or one that is set.

I have not yet ported over the pipeline registration yet, as I'm not
sure if I want to do that inside of the module, or somewhere else.