Port over stemmer
This ports over the stemmer file from elasticlunr.js. I've removed the
self-calling function that was in the original module and just have the
variables defined inside the module. If it turns out that during use I
need the calling function, I'll add it back. I've also not ported the
registration to the pipeline, which will likely happen in a different
location when I add the pipeline.

Most of the code for the stemmer is actually just a copy of a js stemmer
from http://tartarus.org/~martin/PorterStemmer/js.txt