Port over EventEmitter
This is probably a class that I did the most of my own changes to by
far. Instead of it having a remove listener function, I'm just returning
a handle to allow the caller more flexibility to remove handlers, and
I'm using a Map internally instead of just an object. The overall
functionality I think is mostly the same, but with some of my own ideas
(that I stole from other places) in there as well.