Add eslint json config file
After some working through in the codebase and looking at the different
rules and what I want the code to look like, this seems to look ok.

I might add a few more things later (airbnb maybe?) but for now this
seems to give me code that I'm fairly happy with.