Fix linting problems
This is why you want linting from the beginning, so you don't have a
commit like this that just destroys your git history.

Oh well, at least this is fairly early on in the project and it's not
like there's a ton of history for each file.

But this does what it says, it fixes the linting for all of these files
so that everything passes when I run linting.