Port InvertedIndex
I am not 100% sure how this is used yet, but I'm sure I will as I get
used to this library and what's going on.

But from what I can tell, this breaks down tokens into their different
characters and stores them in an object.

The main things I changed were the addToken function taking in a
possible "root" as this was never used in the original code ever, and I
wasn't sure how someone would use that. Maybe later I'll add it back.

I also updated all the refs to just be strings instead of occasionally
numbers, because they aren't stored that way in the object anyways, so
it's easier to just have them be strings all over the place.