Rewrite stemmer
This rewrites the original stemming code using snowball stemming
algorithm.

It's not the cleanest yet, but it's functional. I will be making it
cleaner in later commits when I come back to this.

I chose the snowball stemming algorithm because it seemed like an
improvement on the original porter algorithm, and it was the one I've
seen other similar code use.